podcast/tools/requirements.txt
openai-whisper
anthropic
openai
pillow
